Give each scrape config its own memory
Because everything in the config structs is a pointer we need to take
care that they don't all point to the same global variable.  It seems
neatest to do this by introducing two functions whose return value
escapes.

If they all point to the same place then whatever value is read last
by `UnmarshalYAML()` will apply to all of them.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>