pkg/storage: Fix series_iterator_root by not reusing iterator
It seems that reusing the NewMultiChunkIterator was the problem and
caused the number of read chunks to not be properly tracked. There was
no unit test so I've also created one (that reproduced the issue once I
found it while debugging) and should prevent the same bug happening in
the future.