Faster shutdown
Moves the scrape manager to the top of the run group; so that it gets
interrupted first.

When it got shutdown after the server; the server would gracefully
wait for all connections to be stopped, and since the scrape manager
could be scraping itself it had to wait for all those connections to be
terminated. This terminates them first before shutting down the server.
This results in a significantly faster shutdown time