config: Validate timeout and interval of the scrape configuration
In general we want to stick to the configured interval and don't
overrun because of slow or timing out scraping.

Go's /debug/pprof/profile endpoint will start profiling when being
invoked, has no support for concurrency and the timeout can only
be configured in seconds. The existing code alternated between using
"Seconds", "Delta"+ScrapeInterval but none of these work correctly
when enforcing timeout <= interval.

Make sure that "seconds" < ScrapeTimeout <= ScrapeInterval and adjust
the default configuration to honor this. In the long run we should
consider treating this endpoint entirely different to the others (e.g.
jitter scraping instead of sticking to interval, profile for a longer
duration).

Signed-off-by: Holger Hans Peter Freyther <holger@moiji-mobile.com>