area/ui: Unify value formatting
This PR unifies the two separate mechanisms we use to format numbers right now. The single function for formatting numbers can now be found in `@parca/functions` and currently allows for formatting values in bytes, count, and time.

This also fixes the CPU nanoseconds value not being formatted when hovering over an icicle graph item.

Resolves #409