area/ui: Add additional Next.js pages to the binary pkg
This is a proof of concept for how we can we can create additional Next.js pages and embed them in the built binary package.

After the page has been created in the Next.js app and exported using `make ui`, there should be a directory created for the page with a corresponding `index.html` file.

Finally, the path to the page's directory is added to the embed directive in `ui.go` and the `pkg/server/server.go` file.