Run `podman build` with as many workers as CPUs
To try to make the build a bit faster. Right now it's taking ~45 minutes.

Ideally we would run builds in parallel, too but it seems like support
for this was recently merged (https://github.com/containers/podman/pull/13404), so let's
wait a bit for it to be stable and then we can add it as well if.