actions: Fix build and push job for arm64
To run arm64 arch, QEMU is used. `qemu binfmt-support` packages are
required qemu dependency.

While building container images for arm64 the network is slow due to
container in container mode and yarn install fails if the packages are
big as default retry timeourt is 30sec. Increasing it fixes the issue.

Signed-off-by: Kautilya Tripathi <tripathi.kautilya@gmail.com>