pkg/symbol: Add test for the DWARF symbolizer
While debugging some symbolization issue I wrote this small test to
isolate the problem I was trying to find. While in the end the DWARF
symbolizer behaved correctly, it would be good to have a test to ensure
that things work as expected in the future.

The test executable was copied from
https://github.com/parca-dev/testdata/commit/1636ced7949f341414c116fcdbd17767a2a8e574
and added to this repo because it's quite small (25KB) and will be less
problematic than adding a git submodule.

This commit also:
- changes the demangler in the symbol table symbolizer to be a pointer
- fixes a spelling mistake

```
$ make go/test
```

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>