pkg/debuginfo: Cache positive exists lookups
Previously we issued lots of repeated lookups in object storage. Since
this data is immutable, once this result is positive, it will never
change again, so we can cache the response forever.

This patch caches 10k positive lookups, and also changes the actual API
call issued to be a cheaper one on some platforms (get/head instead of
iter).