deps: update parquet and frostdb (#2570)
* deps: update parquet to 1d85e8136681

This update pulls in null sorting. This update is not necessary for correctness
since sorting is performed through a FrostDB API (that has its own parquet
version), but it's important enough to ingestion that we're better off just
doing this minor bump and forgetting about it.

* deps: update frostdb to d16391d

This pulls in fixes to null sorting order and comparison fixes.