Trim relative to node's parent
So far we have trimmed nodes based on the absolute cumulative value. This is not what we want to do, given we might lose data when diving further into a flame graph.
However, in relation to the parent's cumulative value, there may occasion where the direct child still has to few values and would never be rendered. Those we want to trim.