Trim flame graph nodes relative to node's parent (#2815)
* Trim relative to node's parent

So far we have trimmed nodes based on the absolute cumulative value. This is not what we want to do, given we might lose data when diving further into a flame graph.
However, in relation to the parent's cumulative value, there may occasion where the direct child still has to few values and would never be rendered. Those we want to trim.

* [pre-commit.ci lite] apply automatic fixes

* ui/ProfileIcicleGraph: console.info how many trimmed values

* [pre-commit.ci lite] apply automatic fixes

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>