Fix string trimming table bugs (#2831)
* Fix string trimming table bugs

* pkg/query: Use FlamegraphIterator to walk the graph for trimming

* pkg/query: Don't copy meta data, instead set to empty string

* pkg/query: Remove IDs for aggregated flame graph already

This was actually a regression from a previous PR which I was willing to take for the time being. With the new approach of trimming we don't need to reference anything by ID anymore, and they can be removed earlier.

* pkg/query: Trim locations and functions by setting them to nil

Even though we now keep the locations and functions around, we can still find the once we don't reference anymore and set them to `nil` to save bandwidth.