ui/package/profile/ProfileIcicleGraph: Add more actions
We want to add sorting to the icicle graph itself. This allows to not only sort by function name (the default and what's currently supported) but also sort by cumulative values (such that the highest functions are always left) and in compare mode by diff value (meaning the functions that got worse are left and better right).

This can make finding specific things even easier.