flamegraph_arrow: Inject artifical label node (#3516)
* pkg/query/flamegraph_arrow: Inject pprof label nodes under the root

When grouping by pprof labels we want to inject a node beneath the root into the flamegraph that contains the name of that label. This will make finding the correct stacks by pprof labels easier.

* ui/packages/profile: Parse JSON for label nodes in first level

Instead of showing ``{"thread_name":"foo"}` we're going to parse the JSON and display it as `thread_name="foo"` in a Prometheus fashion.

* pkg/query/flamegraph_arrow: Improve unsafeString usage and simplify isRoot