pkg/query: Use maps instead of slices for children for O(1) access
This improves the performance of building flamegraph API responses
dramatically. The major change is that instead of keeping children as
slices and iterating over them to find which child to step into when
building the graph, we build a key based on the fields we're aggregating
by and use that to check if the children map already contains it.

```
$ benchstat old.txt new.txt
name                old time/op  new time/op  delta
ArrowFlamegraph-10  5.04ms ± 2%  1.58ms ± 2%  -68.66%  (p=0.008 n=5+5)
```