config: Validate duplicate job names
This commit fixes the problem when jobs with the same names are compacted to a single target, leading to an unexpected behavior: a user will only see the last target with this job name.

It is a breaking change for users who have duplicates in their configs, but for new users it will be more transparent.