*: Display CPU cores and CPU seconds for samples count profiles
This patch changes the UX of interacting with CPU profiling data that
only reports samples count. It ensures that samples count is multiplied
by their period (the time between samples), which allows comparing two
profiles that were taken at different frequencies.

As a side effect while changing the queries, this also fixed an issue
where the average cpu time was previously summed up over a step period,
which was incorrect, as the total cpu time over the step needs to be
divided by the total duration over the step.