pkg/config: support seconds in pprof configuration
this updates the definition of `PprofProfilingConfig` such that a new
integer field `Seconds` is added which can be set for profiles.

additionally, it changes the behavior where setting any profiling
configurations will end up overriding the default so merge will now
become a replace.

seconds is accepted as a query parameter and with this change, when
provided, it will be used as scraping duration rather than the scraping
interval value. this allows having scraping where process being profiled
isn't under constant profiling. this only applies to configurations that
have delta enabled.

https://pkg.go.dev/net/http/pprof