Query valuePerSecond and diffPerSecond for delta profiles
For delta profiles (cpu with nanoseconds) we query the value per second, which basically is the `sum(value) / sum(duration)` per stack trace.

We already show this value in the metrics graphs. Now, the icicle graph is going to show it too. More importantly, we're going to use this for diffs.

Instead of sorting by the raw diff value, we now sort by the diff ratio. Basically by the percentage more or less. It works a lot better, especially