*: Store stacktraces in frostdb & symbolize at read time
Storing stacktraces in frostdb has the advantage that the metastore does
not have to be persistent, and instead can be re-built at any point in
time.

This patch is the start of moving to a model, where the metastore is
ephemeral and is only a read-through cache for symbolization, that
happens at query time, as opposed to all the time asynchronously.