fix(snap): store profiles in SNAP_COMMON by default
Until this commit, the default storage location for profiles
on disk was in `$SNAP_DATA/profiles`, but as highlighted in
both #4708 and in https://snapcraft.io/docs/data-locations,
the `$SNAP_DATA` directory is versioned per-revision of the
snap, whereas `$SNAP_COMMON` persists across revisions.

This change leaves config in the same location in `$SNAP_DATA`,
because versioning config files with revisions might still be
desirable.

Existing installs should not be affected, as the default is set
on first run.

Fixes #4708