refactor: Scrape loop. (#5945)
Previously when the scrape suceeded but the function was unable to parse
the result 'continue' was called which made it immediately try again
since the ticker is only waited upon at the end of the loop. This can
cause a huge rps spike if the endpoint is always returning invalid
profiling data.

This moves the processing of the response code to a function instead
that returns an error, and if an error is encountered it handles it and
drops into the second select in the loop to prevent rapid reties.